+++ tests/StackExchange.Redis.Tests/Issues/SO24807536.cs
@@ -20,7 +20,7 @@ namespace StackExchange.Redis.Tests.Issues
                 // setup some data
                 cache.KeyDelete(key, CommandFlags.FireAndForget);
                 cache.HashSet(key, "full", "some value", flags: CommandFlags.FireAndForget);
-                cache.KeyExpire(key, TimeSpan.FromSeconds(1), CommandFlags.FireAndForget);
+                cache.KeyExpire(key, TimeSpan.FromSeconds(2), CommandFlags.FireAndForget);
 
                 // test while exists
                 var keyExists = cache.KeyExists(key);
@@ -31,7 +31,7 @@ namespace StackExchange.Redis.Tests.Issues
                 Assert.Equal("some value", fullWait.Result);
 
                 // wait for expiry
-                await Task.Delay(2000).ForAwait();
+                await UntilCondition(TimeSpan.FromSeconds(10), () => !cache.KeyExists(key)).ForAwait();
 
                 // test once expired
                 keyExists = cache.KeyExists(key);